move hooks definitions to hooks directory
