+++ packages/network-explorer/src/ModelTable.tsx
@@ -1,6 +1,6 @@
 import { Navigate, useParams } from "react-router-dom"
 import { Table } from "./Table.js"
-import { useApplicationData } from "./useApplicationData.js"
+import { useApplicationData } from "./hooks/useApplicationData.js"
 
 export const ModelTable = ({
 	showSidebar,
