+++ packages/network-explorer/src/Sidebar.tsx
@@ -4,7 +4,7 @@ import { TableSelector } from "./TableSelector.js"
 import { LuTable2 } from "react-icons/lu"
 import { ApplicationData } from "./ApplicationData.js"
 import { TableDef } from "./tables.js"
-import { useApplicationData } from "./useApplicationData.js"
+import { useApplicationData } from "./hooks/useApplicationData.js"
 import { useState } from "react"
 
 export const Sidebar = ({ tables }: { tables: TableDef[] }) => {
